deleted functionality folder; no longer need to run child-process outside, can add to switch statement
